Board and Collision Logic moved to Logic Layer.
